Rename IpLoc to IpLocManager
