+++ baremaps-core/src/main/java/org/apache/baremaps/iploc/IpLocRepository.java
@@ -10,7 +10,7 @@
  * the License.
  */
 
-package org.apache.baremaps.iploc.database;
+package org.apache.baremaps.iploc;
